move registration summary full into shared module
