+++ embc-app/ClientApp/src/app/shared/components/registration-summary-full/registration-summary-full.component.ts
